Convert an np.float64 back to a float.
Most of these were changed in
Author: Jackson Burns <jwburns@mit.edu>
Date:   Sat Jul 22 15:35:13 2023 -0400
    fix use of np.float and np.bool
but this one got missed or switched back somehow.

+++ arkane/statmech.py
@@ -1111,7 +1111,7 @@ def project_rotors(conformer, hessian, rotors, linear, is_ts, get_projected_out_
             i += 1
 
     # T is the transformation vector from cartesian to internal coordinates
-    T = np.zeros((n_atoms * 3, 3 * n_atoms - external), np.float64)
+    T = np.zeros((n_atoms * 3, 3 * n_atoms - external), float)
 
     T[:, 0:3 * n_atoms - external] = p[:, external:3 * n_atoms]
